requirements.txt
fastapi
uvicorn
pipecat-ai[twilio,google,silero]
python-dotenv
pydantic-settings
twilio
